remove nhib cfg to web cfg
