--- NewsWebSite/Models/NHibernate/Nhibernate.cfg.xml
@@ -1,17 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<hibernate-configuration xmlns="urn:nhibernate-configuration-2.2">
-  <session-factory>
-    <property name="connection.provider">
-      NHibernate.Connection.DriverConnectionProvider
-    </property>
-    <property name="connection.driver_class">
-      NHibernate.Driver.SqlClientDriver
-    </property>
-    <property name="connection.connection_string">
-      Server=DIMAS\SQLEXPRESS; database=NewsWebSiteDB; Integrated Security=False; User Id=root; Password=root; MultipleActiveResultSets=True;
-    </property>
-    <property name="dialect">
-      NHibernate.Dialect.MsSql2012Dialect
-    </property>
-  </session-factory>
-</hibernate-configuration>
