+++ NewsWebSite/Models/Repository/NHibernateRepository.cs
@@ -1,26 +1,13 @@
-﻿using NewsWebSite.Controllers;
-using NewsWebSite.Models.ViewModel;
-using Newtonsoft.Json;
-using NHibernate;
+﻿using NHibernate;
 using NHibernate.Criterion;
 using NHibernate.Transform;
 using System;
 using System.Collections.Generic;
-using System.Configuration;
 using System.Linq;
 using System.Web;
 
-namespace NewsWebSite.Models
+namespace NewsWebSite.Models.Repository
 {
-    public interface IRepository
-    {
-        Article GetItem(int id);
-        int GetCountOfLines();
-        PagedList<Article> GetList(int starFrom, int count);
-        PagedList<DemoArticle> GetDemoList(int starFrom, int count);
-        int Save(Article article);
-    }
-
     public class NHibernateRepository : IRepository
     {
         readonly ISessionFactory sessionFactory;
@@ -30,7 +17,7 @@ namespace NewsWebSite.Models
         {
             this.sessionFactory = sessionFactory;
         }
-      
+
 
         public int Save(Article a)
         {
@@ -65,13 +52,13 @@ namespace NewsWebSite.Models
             using (var session = sessionFactory.OpenSession())
             {
                 var results = new PagedList<Article>();
-                    results.AddRange(session.CreateCriteria<Article>()
-                    .SetFirstResult(starFrom)
-                    .SetMaxResults(count)
-                    .AddOrder(Order.Desc("Id"))
-                    .List<Article>());
+                results.AddRange(session.CreateCriteria<Article>()
+                .SetFirstResult(starFrom)
+                .SetMaxResults(count)
+                .AddOrder(Order.Desc("Id"))
+                .List<Article>());
                 results.LinesCount = session.QueryOver<Article>().Select(Projections.RowCount()).FutureValue<int>().Value;
-                results.PageCount = (int)Math.Ceiling(results.LinesCount / double.Parse(ConfigurationManager.AppSettings["NumberOfItemsOnPage"]));
+                results.PageCount = (int)Math.Ceiling(results.LinesCount / double.Parse(System.Configuration.ConfigurationManager.AppSettings["NumberOfItemsOnPage"]));
                 return results;
             }
         }
@@ -96,7 +83,7 @@ namespace NewsWebSite.Models
                     .SetResultTransformer(Transformers.AliasToBean<DemoArticle>())
                     .List<DemoArticle>());
                 results.LinesCount = session.QueryOver<Article>().Select(Projections.RowCount()).FutureValue<int>().Value;
-                results.PageCount = (int)Math.Ceiling(results.LinesCount / double.Parse(ConfigurationManager.AppSettings["NumberOfItemsOnPage"]));
+                results.PageCount = (int)Math.Ceiling(results.LinesCount / double.Parse(System.Configuration.ConfigurationManager.AppSettings["NumberOfItemsOnPage"]));
                 return results;
             }
         }
@@ -110,41 +97,5 @@ namespace NewsWebSite.Models
                 return count;
             }
         }
-    }
-    
-    /* public class CachedRepository : IRepository
-     {
-         readonly IRepository _realRepo;
-
-         public CachedRepository(IRepository realRepo)
-         {
-             _realRepo = realRepo;
-         }
-
-         public Article GetItem(int id)
-         {
-
-             //check if article exsit in cache
-             var article = _realRepo.GetItem(id);
-
-             var cacheKey = string.Join("::", id, article.LastUpdateDate);
-             //put to cache
-         }
-
-         public PagedList<Article> GetList(int offset = 0, int count = 10)
-         {
-             throw new NotImplementedException();
-         }
-
-         public void Save(Article article)
-         {
-             //remove from cache by key
-         }
-     } */
-
-    public class PagedList<T> : List<T>
-    {
-        public int LinesCount { get; set; }
-        public int PageCount { get; set; }
     }
 }
