+++ NewsWebSite/NewsWebSite.csproj
@@ -236,12 +236,12 @@
     <Compile Include="Models\ArticleForView.cs" />
     <Compile Include="Models\DemoArticle.cs" />
     <Compile Include="Models\IdentityModels.cs" />
-    <Compile Include="Models\IRepository.cs" />
+    <Compile Include="Models\Repository\NHibernateRepository.cs" />
+    <Compile Include="Models\Repository\IRepository.cs" />
     <Compile Include="Models\ManageViewModels.cs" />
     <Compile Include="Models\ViewModel\CreateArticleModel.cs" />
     <Compile Include="Models\ViewModel\EditArticleModel.cs" />
     <Compile Include="Models\ViewModel\ListItemPageModel.cs" />
-    <Compile Include="NHibernateHelper.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Startup.cs" />
   </ItemGroup>
@@ -256,9 +256,6 @@
     <Content Include="Global.asax" />
     <Content Include="Content\Site.css" />
     <EmbeddedResource Include="Models\NHibernate\Article.hbm.xml" />
-    <Content Include="Models\NHibernate\Nhibernate.cfg.xml">
-      <SubType>Designer</SubType>
-    </Content>
     <Content Include="Scripts\ai.0.22.19-build00125.js" />
     <Content Include="Scripts\ai.0.22.19-build00125.min.js" />
     <Content Include="Scripts\bootstrap.js" />
@@ -288,7 +285,6 @@
     <Content Include="Scripts\respond.js" />
     <Content Include="Scripts\respond.min.js" />
     <Content Include="Scripts\_references.js" />
-    <Content Include="Content\UserImages\3177\f-sVdMj24c0.jpg" />
     <Content Include="Web.config" />
     <Content Include="Web.Debug.config">
       <DependentUpon>Web.config</DependentUpon>
@@ -331,6 +327,7 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="App_Data\" />
+    <Folder Include="Content\UserImages\" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="fonts\glyphicons-halflings-regular.woff" />
