+++ NewsWebSite/Web.config
@@ -5,14 +5,32 @@
   -->
 <configuration>
   <configSections>
+    <section name="hibernate-configuration" type="NHibernate.Cfg.ConfigurationSectionHandler, NHibernate" />
     <!-- For more information on Entity Framework configuration, visit http://go.microsoft.com/fwlink/?LinkID=237468 -->
     <section name="entityFramework" type="System.Data.Entity.Internal.ConfigFile.EntityFrameworkSection, EntityFramework, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false" />
   </configSections>
+
+  <hibernate-configuration xmlns="urn:nhibernate-configuration-2.2">
+    <session-factory>
+      <property name="connection.provider">
+        NHibernate.Connection.DriverConnectionProvider
+      </property>
+      <property name="connection.driver_class">
+        NHibernate.Driver.SqlClientDriver
+      </property>
+      <property name="connection.connection_string">
+        Server=DIMAS\SQLEXPRESS; database=NewsWebSiteDB; Integrated Security=False; User Id=root; Password=root; MultipleActiveResultSets=True;
+      </property>
+      <property name="dialect">
+        NHibernate.Dialect.MsSql2012Dialect
+      </property>
+    </session-factory>
+  </hibernate-configuration>
+
   <connectionStrings>
     <add name="DefaultConnection" connectionString="Data Source=(LocalDb)\MSSQLLocalDB;AttachDbFilename=|DataDirectory|\aspnet-NewsWebSite-20170115063838.mdf;Initial Catalog=aspnet-NewsWebSite-20170115063838;Integrated Security=True" providerName="System.Data.SqlClient" />
   </connectionStrings>
   <appSettings>
-    <add key="NHibernateConfig" value="~\Models\NHibernate\Nhibernate.cfg.xml"/>
     <add key="UserImagesFolder" value="~/Content/UserImages/" />
     <add key ="NumberOfItemsOnPage" value="15" />
     <add key="webpages:Version" value="3.0.0.0" />
